include a projection test for validation
the reprojection error seems to be quite large, we need 8px for the test
to pass by comparing the stored pixel coordinates with the computed
coordinates using ImageProjectiGeometry. As we are not testing the
projection, but loading the file, I guess this is fine. But still keep
this in mind...